refactor(common) Generalizes visitor from commonmark
